requirements.txt
easyocr==1.7.1
opencv-python==4.8.1.78
numpy==1.24.3
torch==2.1.0
torchvision==0.16.0
matplotlib==3.7.2
pyyaml==6.0.1
tqdm==4.66.1
pillow==10.0.0
scikit-learn==1.3.0
flask==3.0.0 